feat: implement route blueprints for teams, players, matches, news, and results
